requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
Deprecated==1.2.13
Django==4.1.7
django-cors-headers==3.14.0
django-filter==23.1
django-oauth-toolkit==2.2.0
django-otp==1.1.6
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.2
oauth2-provider==0.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pycparser==2.21
pytz==2023.3
requests==2.29.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
wrapt==1.15.0
